Added Travis CI intergration
As with my SRFI 113 and 114, the installation is sorta dumb. Too busy
developing these libraries to take time to make Snow2 usable for CI.

Installation is going to become even more complex for this library
because of the native code, as currently Snow2 cannot install such
libraries in *any* way.