Unroll list->integer by 16 bits
Most of lagging here came from lots of temporary bignums which result
from adding bits one by one. Adding bits in chunks of 16 obviously
reduces the number of temporary values by 16.

Why sixteen? Because we are sitting on a fence here. We want the buffer
to remain a fixnum for quick accumulation, and also to avoid 'hard-to-do'
shifts in bignums (read: non-power-of-two). 64-bit machines have 61-bit
fixnums and 64-bit limbs of bignums; 32-bit machines have 29-bit fixnums
and 32-bit limbs. 64-bit accumulator can be promoted to a bignum on both
architectures, 32-bit works well for 64 bits, but can be subpar on 32-bit
machines, thus we take 16 bits as the highest suitable power of two.