Inline and constant-fold bitops
Frankly speaking, this optimization should be performed by the
implementation, but Chibi is just too stupid to do it properly.

There is no inlining of procedure calls at all, so all bit operations
are hand-inlined via macros. This brings back bitwise-not definition.

Chibi can constant fold arithmetics, but does not do this for `expt`
(as it is a library call). Instead of tweaking the implementation we
will just inline fixed-argument shifts and masks by ourselves.

All this yields us around 33% more performance on Chibi.