Iteration over bignum limbs in two's-complement
Added a function for efficient iteration over bignum's limbs with
on-the-fly conversion to equivalent two's-complement representation.

Example usage:

    void print_bignum_limbs(sexp num)
    {
        sexp_uint_t i;
        sexp_uint_t len = sexp_bignum_length(num);
        int state = START_ITERATION;

        for (i = 0; i < len; i++)
            printf("0x%016llX ", bignum_next_limb(num, i, &state));

        printf("\n");
    }

As you may have noticed, bignums are using sign-magnitude repesentation
which is _slightly unwieldy_ for bitwise arithmetics that assume two's-
complement one. We must do something about this. Consider, for example,
calculating `bit-and` for bignums. One option would be to allocate two
arrays, convert both numbers into their two's-complement representation
placing the results into said arrays, perform in-place conjunction, and
finally convert the result back into sign-magnitude form and throw away
the second array. However, this places unnecessary stress onto garbage
collector. To avoid it we strive to perform the conversion on-the-fly,
without allocating scratch buffers.

For positive numbers the conversion is identity, we only need to check
for overflow: when the index exceeds the significant range of a bignum.
In this case we should return just zeros as it is the sign-extension of
positive numbers.

It is slightly harder for negative ones. We cannot just invert the limb
because two's-complement negation is 'inverse + 1'. This unit addition
can result into a carry if the previous limb is zero:

    -42 => - 0010 1010    -800 => - 1000 0010 1010 0000    (4-bit limbs)
    ... 1111 1101 0110     ... 1111 0111 1101 0110 0000
        ssss iiii nnnn         ssss iiii iiii nnnn nnnn

    where:   s - sign-extend,  i - invert,  n - negate

This is why we use negation until we meet a non-zero limb (including the
very first one). After that only inversion is used (until the end of the
significant bits).

Proper inlining pays off here as the compiler would be able to generate
two separate loops for positive and negative numbers. Also, the state-
transition graph does not contain cycles so all branches should be well-
suited for the branch prediction as well.