Version 1.0.0 release
This version includes a portable implementation of SRFI 60. It is
somewhat optimized in comparison with the reference implementation,
but obviously we cannot make it *real fast* in R7RS Scheme which
does not define any bitwise operations and abstracts out numbers
as much as possible.

I believe that this will do, as it is as close as we can get to a
fast implementation while still staying in the portability domain.
Native implementations are going to be faster (estimated boost is
around factor of 10 for fixnums, and much greater for bignums).